(ref)Move copyText function to web.js file
