+++ react/features/base/util/index.web.js
@@ -1,0 +1,1 @@
+export * from './copyText';
